refactor: google analytics를 shared/lib로 이동
변경 사항:
- util/googleAnalytics/ → shared/lib/google-analytics/로 이동
- shared/lib/index.ts에서 export 추가
- 기능은 동일하게 유지

이유:
- FSD 아키텍처에 맞게 유틸리티를 shared/lib로 통합
- google analytics 관련 코드의 재사용성 향상
