+++ client/src/shared/lib/google-analytics/index.ts
@@ -1,0 +1,4 @@
+export { initializeGA } from './initializeGA';
+export { trackEvent } from './trackEvent';
+export { default as PageViewTracker } from './PageViewTracker';
+
