+++ client/src/shared/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './history';
+export * from './google-analytics';
